requirements.txt
# Stock Data Pipeline - Python Dependencies

# Data Processing
pandas>=2.0.0
pyarrow>=14.0.0
numpy>=1.24.0

# MinIO (S3-compatible storage)
minio>=7.2.0

# Database
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0

# API Requests
requests>=2.31.0

# Apache Airflow (for local development)
apache-airflow>=2.7.0

# Utilities
python-dotenv>=1.0.0

# docker compose up -d --build